complete instructions of nft staking

+++ week-4/anchor-nft-staking-starter-q4-25/programs/anchor-nft-staking-q4-25/src/instructions/initialize_config.rs
@@ -36,7 +36,7 @@ impl<'info> InitializeConfig<'info> {
         &mut self,
         points_per_stake: u8,
         max_stake: u8,
-        freeze_period: u32,
+        freeze_period: u64,
         bumps: &InitializeConfigBumps,
     ) -> Result<()> {
         self.config.set_inner(StakeConfig {
