+++ week-4/anchor-nft-staking-starter-q4-25/programs/anchor-nft-staking-q4-25/src/state/stake_config.rs
@@ -5,7 +5,7 @@ use anchor_lang::prelude::*;
 pub struct StakeConfig {
     pub points_per_stake: u8,
     pub max_stake: u8,
-    pub freeze_period: u32,
+    pub freeze_period: u64,
     pub rewards_bump: u8,
     pub bump: u8,
 }
